move numba modules
